+++ CHANGES.md
@@ -4,6 +4,8 @@ CHANGELOG
 0.4.4.dev0     (XXXX-XX-XX)
 ---------------------------
 
+* add relation property to handle relations directly in feature
+
 
 0.4.3          (2020-10-09)
 ---------------------------